requirements.txt
aiohttp==3.8.4 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.5.7 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
filelock==3.12.0 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2023.5.0 ; python_version >= "3.10" and python_version < "4.0"
gtts==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.14.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
networkx==3.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.3 ; python_version >= "3.10" and python_version < "4.0"
openai==0.27.6 ; python_version >= "3.10" and python_version < "4.0"
opencv-python==4.7.0.72 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.10" and python_version < "4.0"
playsound==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic[dotenv]==1.10.7 ; python_version >= "3.10" and python_version < "4.0"
pytelegrambotapi==4.11.0 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
regex==2023.5.5 ; python_version >= "3.10" and python_version < "4.0"
requests==2.30.0 ; python_version >= "3.10" and python_version < "4.0"
sympy==1.12 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.13.3 ; python_version >= "3.10" and python_version < "4.0"
torch==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
transformers==4.29.1 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.10" and python_version < "4.0"
